check if folder exits before clearing them

+++ testing/PrepareServer.hx
@@ -65,7 +65,9 @@ class PrepareServer {
 	static function main() {
 		Sys.println("Cleaning up folders");
 		var serverPath = Sys.getCwd() + 'server/';
-		deleteFolderContent(serverPath+ 'files');
+		if (FileSystem.exists(serverPath+ 'files') && FileSystem.isDirectory(serverPath+ 'files'))
+			deleteFolderContent(serverPath+ 'files');
+		if (FileSystem.exists(serverPath+ 'tmp') && FileSystem.isDirectory(serverPath+ 'tmp'))
 		deleteFolderContent(serverPath+ 'tmp');
 		
 		Sys.println("Cleaning DB");
